Moved code from io/vi*.c into ultra_reimplementation.cpp. And other code cleanup.
