+++ src/libultra/io/viextend.c
@@ -4,6 +4,3 @@
 
 extern u32 __additional_scanline;
 
-void osViExtendVStart(u32 arg0) {
-    __additional_scanline = arg0;
-}
