+++ src/libultra/io/visetyscale.c
@@ -5,11 +5,3 @@
 
 extern OSViContext* __osViNext;
 
-void osViSetYScale(f32 scale) {
-    //register u32 prevInt = __osDisableInt();
-
-    __osViNext->y.factor = scale;
-    __osViNext->state |= 4;
-
-    //__osRestoreInt(prevInt);
-}
